Applied PR comments
